got rid of old structure but lagacy code is still present
