+++ src/legacy/mod.rs
@@ -1,0 +1,9 @@
+mod budget;
+mod date;
+pub mod goal;
+pub mod input;
+pub mod output;
+mod repetition;
+mod slot;
+mod step;
+mod timeline;
