+++ src/legacy/step.rs
@@ -7,8 +7,6 @@ use super::{
     timeline::Timeline,
 };
 
-pub mod impls;
-
 /// Steps are generated to achieve a Goal in one or more Steps.
 /// A leaf Goal can generate one or more Steps.
 #[derive(Deserialize, Debug, Eq, Clone, PartialEq)]
