--- src/new_models/mod.rs
@@ -1,6 +1,0 @@
-pub mod calendar;
-pub mod date;
-pub mod day;
-pub mod day_filter;
-pub mod flexibility;
-pub mod goal;
